requirements.txt
numpy==1.18.1
torchvision==0.5.0
torch==1.4.0
opencv_python==4.2.0.32
matplotlib==3.1.3
six==1.14.0
Pillow==7.1.1
visdom==0.1.8.9
